[APPC-4447] Removed VK implementation
