+++ app/src/main/java/com/asfoundation/wallet/billing/adyen/PaymentType.java
@@ -13,8 +13,6 @@ public enum PaymentType {
 
   CHALLENGE_REWARD(Collections.singletonList("challenge_reward")),
 
-  VKPAY(Collections.singletonList("vk_pay")),
-
   LOCAL_PAYMENTS(Collections.singletonList("localPayments")),
 
   CARRIER_BILLING(Collections.singletonList("carrier_billing")),
