+++ app/src/main/java/com/asfoundation/wallet/billing/amazonPay/SuccessInfo.kt
@@ -1,0 +1,7 @@
+package com.asfoundation.wallet.billing.amazonPay
+
+data class SuccessInfo(
+  val hash: String?,
+  val orderReference: String?,
+  val purchaseUid: String?,
+)
