+++ app/src/main/java/com/asfoundation/wallet/topup/TopUpFragment.kt
@@ -730,9 +730,6 @@ class TopUpFragment : BasePageViewFragment(), TopUpFragmentView {
         PaymentType.CHALLENGE_REWARD.subTypes.contains(data.id) ->
           PaymentTypeInfo(PaymentType.CHALLENGE_REWARD, data.id, data.label, data.iconUrl)
 
-        PaymentType.VKPAY.subTypes.contains(data.id) ->
-          PaymentTypeInfo(PaymentType.VKPAY, data.id, data.label, data.iconUrl)
-
         PaymentType.GOOGLEPAY_WEB.subTypes.contains(data.id) ->
           PaymentTypeInfo(PaymentType.GOOGLEPAY_WEB, data.id, data.label, data.iconUrl)
