+++ app/src/main/java/com/asfoundation/wallet/ui/iab/IabView.kt
@@ -74,18 +74,6 @@ interface IabView {
     frequency: String?
   )
 
-  fun showVkPay(
-    amount: BigDecimal,
-    currency: String?,
-    isBds: Boolean,
-    paymentType: PaymentType,
-    bonus: String?,
-    iconUrl: String?,
-    gamificationLevel: Int,
-    isSubscription: Boolean,
-    frequency: String?
-  )
-
   fun showAmazonPay(
     amount: BigDecimal,
     currency: String?,
