+++ app/src/main/java/com/asfoundation/wallet/ui/iab/PaymentMethodsAnalytics.kt
@@ -29,7 +29,6 @@ class PaymentMethodsAnalytics @Inject constructor(
     const val PAYMENT_METHOD_CC = "credit_card"
     const val PAYMENT_METHOD_PP = "paypal"
     const val PAYMENT_METHOD_PP_V2 = "paypal_v2"
-    const val PAYMENT_METHOD_VKPAY = "vk_pay"
     const val PAYMENT_METHOD_APPC = "appc_c"
     const val PAYMENT_METHOD_LOCAL = "local"
     const val PAYMENT_METHOD_SANDBOX = "sandbox"
