+++ app/src/main/java/com/asfoundation/wallet/ui/iab/PaymentMethodsFragment.kt
@@ -55,7 +55,6 @@ import io.reactivex.subjects.PublishSubject
 import java.math.BigDecimal
 import javax.inject.Inject
 
-
 @AndroidEntryPoint
 class PaymentMethodsFragment : BasePageViewFragment(), PaymentMethodsView {
 
@@ -710,25 +709,6 @@ class PaymentMethodsFragment : BasePageViewFragment(), PaymentMethodsView {
         frequency
       )
     }
-  }
-
-  override fun showVkPay(
-    gamificationLevel: Int,
-    fiatValue: FiatValue,
-    frequency: String?,
-    isSubscription: Boolean
-  ) {
-    iabView.showVkPay(
-      fiatValue.amount,
-      fiatValue.currency,
-      isBds,
-      PaymentType.PAYPAL,
-      bonusMessageValue,
-      null,
-      gamificationLevel,
-      isSubscription,
-      frequency
-    )
   }
 
   override fun showAmazonPay(
