+++ app/src/main/java/com/asfoundation/wallet/ui/iab/PaymentMethodsView.kt
@@ -64,11 +64,6 @@ interface PaymentMethodsView {
     isSubscription: Boolean
   )
 
-  fun showVkPay(
-    gamificationLevel: Int, fiatValue: FiatValue, frequency: String?,
-    isSubscription: Boolean
-  )
-
   fun showAmazonPay(
     gamificationLevel: Int, fiatValue: FiatValue, frequency: String?,
     isSubscription: Boolean
@@ -167,7 +162,7 @@ interface PaymentMethodsView {
 
   enum class SelectedPaymentMethod {
     PAYPAL, PAYPAL_V2, CREDIT_CARD, APPC, APPC_CREDITS, MERGED_APPC, SHARE_LINK, LOCAL_PAYMENTS,
-    EARN_APPC, CARRIER_BILLING, ERROR, SANDBOX, CHALLENGE_REWARD, VKPAY, GOOGLEPAY_WEB, MI_PAY, AMAZONPAY
+    EARN_APPC, CARRIER_BILLING, ERROR, SANDBOX, CHALLENGE_REWARD, GOOGLEPAY_WEB, MI_PAY, AMAZONPAY
   }
 
   enum class PaymentMethodId(val id: String) {
@@ -181,7 +176,6 @@ interface PaymentMethodsView {
     SANDBOX("sandbox"),
     ASK_FRIEND("ask_friend"),
     CHALLENGE_REWARD("challenge_reward"),
-    VKPAY("vk_pay"),
     GOOGLEPAY_WEB("googlepay"),
     MI_PAY("mipay"),
     AMAZONPAY("amazonpay")
