--- app/src/main/res/layout/onboarding_vk_payment_layout.xml
@@ -1,118 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/onboarding_vk_payment_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/styleguide_blue_secondary"
-    >
-
-  <com.vk.auth.ui.fastloginbutton.VkFastLoginButton
-      android:id="@+id/vk_fast_login_button"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="30dp"
-      android:visibility="gone"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-
-  <com.airbnb.lottie.LottieAnimationView
-      android:id="@+id/loading_authorization_animation"
-      android:layout_width="144dp"
-      android:layout_height="144dp"
-      android:layout_centerInParent="true"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      app:layout_constraintVertical_bias="0.41000003"
-      app:lottie_autoPlay="true"
-      app:lottie_enableMergePathsForKitKatAndAbove="true"
-      app:lottie_loop="true"
-      app:lottie_rawRes="@raw/loading_wallet"
-      tools:visibility="visible"
-      />
-
-  <ScrollView
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      >
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        >
-
-      <FrameLayout
-          android:id="@+id/complete_payment_view"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:layout_marginBottom="32dp"
-          android:background="@color/styleguide_blue_secondary"
-          android:visibility="gone"
-          app:layout_constraintBottom_toTopOf="@+id/onboarding_success_vk_buttons"
-          app:layout_constraintEnd_toEndOf="parent"
-          app:layout_constraintStart_toStartOf="parent"
-          app:layout_constraintTop_toTopOf="parent"
-          app:layout_constraintVertical_bias="0.36"
-          >
-        <include
-            android:id="@+id/fragment_first_iab_transaction_completed"
-            layout="@layout/fragment_first_iab_transaction_completed"
-            />
-      </FrameLayout>
-
-      <include
-          android:id="@+id/onboarding_success_vk_buttons"
-          layout="@layout/onboarding_payment_success_buttons"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:visibility="gone"
-          app:layout_constraintBottom_toBottomOf="parent"
-          app:layout_constraintEnd_toEndOf="parent"
-          app:layout_constraintStart_toStartOf="parent"
-          tools:visibility="visible"
-          />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-  </ScrollView>
-
-  <include
-      android:id="@+id/no_network"
-      layout="@layout/no_network_retry_only_layout"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:visibility="gone"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-
-  <include
-      android:id="@+id/error_view"
-      layout="@layout/error_top_up_layout"
-      android:layout_width="0dp"
-      android:layout_height="0dp"
-      android:visibility="gone"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-
-  <com.appcoins.wallet.ui.widgets.WalletButtonView
-      android:id="@+id/error_try_again_vk"
-      android:layout_width="wrap_content"
-      android:layout_height="@dimen/small_button_height"
-      android:layout_marginTop="24dp"
-      android:visibility="gone"
-      app:buttonText="@string/try_again"
-      app:buttonType="filled"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@id/error_view"
-      />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
