--- app/src/main/res/layout/vk_payment_iab_layout.xml
@@ -1,119 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/background_card"
-    >
-  <androidx.constraintlayout.widget.ConstraintLayout
-      android:id="@+id/main_content"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      app:layout_constraintTop_toTopOf="parent"
-      >
-
-    <com.vk.auth.ui.fastloginbutton.VkFastLoginButton
-        android:id="@+id/vk_fast_login_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="30dp"
-        android:visibility="gone"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        />
-
-  </androidx.constraintlayout.widget.ConstraintLayout>
-
-  <com.airbnb.lottie.LottieAnimationView
-      android:id="@+id/loading"
-      android:layout_width="128dp"
-      android:layout_height="128dp"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      app:lottie_autoPlay="true"
-      app:lottie_enableMergePathsForKitKatAndAbove="true"
-      app:lottie_loop="true"
-      app:lottie_rawRes="@raw/loading_wallet"
-      android:visibility="invisible"
-      tools:visibility="visible"
-      />
-
-  <TextView
-      android:id="@+id/loading_hint_text_view"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:text="@string/purchase_making_purchase_title"
-      android:visibility="invisible"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@id/loading"
-      />
-
-  <include
-      android:id="@+id/success_container_vk"
-      layout="@layout/fragment_iab_transaction_completed"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      app:layout_constraintVertical_bias="0.32"
-      />
-
-  <include
-      android:id="@+id/no_network"
-      layout="@layout/no_network_retry_only_layout"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:visibility="gone"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-
-  <include
-      android:id="@+id/error_view"
-      layout="@layout/support_error_layout"
-      android:layout_width="match_parent"
-      android:layout_height="0dp"
-      android:visibility="gone"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-
-
-  <com.appcoins.wallet.ui.widgets.WalletButtonView
-      android:id="@+id/error_cancel_vk"
-      android:layout_width="wrap_content"
-      android:layout_height="@dimen/small_button_height"
-      android:layout_marginEnd="16dp"
-      android:layout_marginBottom="16dp"
-      android:visibility="gone"
-      app:buttonColor="@color/styleguide_dark_grey"
-      app:buttonText="@string/cancel_button"
-      app:buttonType="text"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toStartOf="@id/error_try_again_vk"
-      />
-
-  <com.appcoins.wallet.ui.widgets.WalletButtonView
-      android:id="@+id/error_try_again_vk"
-      android:layout_width="wrap_content"
-      android:layout_height="@dimen/small_button_height"
-      android:layout_marginEnd="16dp"
-      android:layout_marginBottom="16dp"
-      android:visibility="gone"
-      app:buttonText="@string/try_again"
-      app:buttonType="filled"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
