+++ app/src/main/res/navigation/inner_payment_graph.xml
@@ -38,10 +38,6 @@
         app:destination="@id/onboarding_local_payment_fragment" />
 
     <action
-        android:id="@+id/action_navigate_to_onboarding_vk_payment"
-        app:destination="@id/onboarding_vk_payment_fragment" />
-
-    <action
         android:id="@+id/action_navigate_to_onboarding_google_pay"
         app:destination="@id/onboarding_google_pay_fragment" />
 
@@ -135,33 +131,6 @@
     <argument
         android:name="currency"
         app:argType="string" />
-
-    <action
-        android:id="@+id/action_navigate_to_onboarding_payment_result"
-        app:destination="@id/onboarding_payment_result_fragment" />
-  </fragment>
-
-  <fragment
-      android:id="@+id/onboarding_vk_payment_fragment"
-      android:name="com.asfoundation.wallet.onboarding_new_payment.vkPayment.OnboardingVkPaymentFragment"
-      tools:layout="@layout/onboarding_vk_payment_layout">
-
-    <argument
-        android:name="transaction_builder"
-        app:argType="com.asfoundation.wallet.entity.TransactionBuilder" />
-
-    <argument
-        android:name="payment_type"
-        app:argType="com.asfoundation.wallet.billing.adyen.PaymentType" />
-    <argument
-        android:name="amount"
-        app:argType="string" />
-    <argument
-        android:name="currency"
-        app:argType="string" />
-    <argument
-        android:name="forecast_bonus"
-        app:argType="com.appcoins.wallet.gamification.repository.ForecastBonusAndLevel" />
 
     <action
         android:id="@+id/action_navigate_to_onboarding_payment_result"
