+++ core/analytics/src/main/java/com/appcoins/wallet/core/analytics/analytics/legacy/BillingAnalytics.kt
@@ -505,7 +505,6 @@ class BillingAnalytics @Inject constructor(
     const val PAYMENT_METHOD_PAYPAL = "PAYPAL"
     const val PAYMENT_METHOD_PAYPALV2 = "PAYPAL_V2"
     const val PAYMENT_METHOD_GOOGLE_PAY_WEB = "GOOGLE_PAY"
-    const val PAYMENT_METHOD_VK_PAY = "VK_PAY"
     const val PAYMENT_METHOD_CARRIER = "CARRIER"
     const val PAYMENT_METHOD_SANDBOX = "SANDBOX"
     const val PAYMENT_METHOD_MI_PAY = "MI_PAY"
