--- core/network/microservices/src/main/java/com/appcoins/wallet/core/network/microservices/model/VkPayPaymentRequest.kt
@@ -1,29 +1,0 @@
-package com.appcoins.wallet.core.network.microservices.model
-
-import com.google.gson.annotations.SerializedName
-
-
-data class VkPayPaymentRequest(
-  @SerializedName("callback_url") val callbackUrl: String?,
-  @SerializedName("domain") val domain: String?,
-  @SerializedName("metadata") val metadata: String?,
-  @SerializedName("origin") val origin: String?,
-  @SerializedName("product") val sku: String?,
-  @SerializedName("reference") val reference: String?,
-  @SerializedName("type") val type: String?,
-  @SerializedName("price") val price: VkPrice,
-  @SerializedName("entity.oemid") val entityOemId: String?,
-  @SerializedName("entity.domain") val entityDomain: String?,
-  @SerializedName("entity.promo_code") val entityPromoCode: String?,
-  @SerializedName("wallets.user") val user: String?,
-  @SerializedName("referrer_url") val referrerUrl: String?,
-  @SerializedName("entity.guest_id") val guestWalletId: String?,
-  @SerializedName("method") val method: String?,
-  @SerializedName("email") val email: String?,
-  @SerializedName("phone_number") val phone: String?,
-)
-
-data class VkPrice(
-  val currency: String,
-  val value: String
-)
