--- core/network/microservices/src/main/java/com/appcoins/wallet/core/network/microservices/model/VkPayTransaction.kt
@@ -1,11 +1,0 @@
-package com.appcoins.wallet.core.network.microservices.model
-
-data class VkPayTransaction(
-  val uid: String?,
-  val hash: String?,
-  val status: TransactionStatus?,
-  val amount: Int?,
-  val merchantId: String,
-  val errorCode: String? = null,
-  val errorMessage: String? = null,
-)
