--- core/network/microservices/src/main/java/com/appcoins/wallet/core/network/microservices/model/VkTransactionResponse.kt
@@ -1,11 +1,0 @@
-package com.appcoins.wallet.core.network.microservices.model
-
-import com.google.gson.annotations.SerializedName
-
-data class VkTransactionResponse(
-  val uid: String, val hash: String?,
-  @SerializedName("reference") val orderReference: String?,
-  val status: TransactionStatus,
-  val amount: Int?,
-  @SerializedName("merchant_id") val merchantId: String?,
-)
