--- feature/vk-pay/build.gradle.kts
@@ -1,11 +1,0 @@
-plugins {
-  id("appcoins.android.library.compose")
-}
-
-android {
-  namespace = "com.appcoins.wallet.feature.vkpay"
-}
-
-dependencies {
-  implementation(libs.bundles.vk)
-}
