--- feature/vk-pay/src/main/java/com/appcoins/wallet/feature/vkpay/VKConstants.kt
@@ -1,9 +1,0 @@
-package com.appcoins.wallet.feature.vkpay
-
-object VKConstants {
-
-  const val APP_VK_VERSION = "1.232"
-  internal const val SERVICE_USER_AGREEMENT = "https://id.vk.com/terms"
-  internal const val SERVICE_PRIVACY_POLICY = "https://id.vk.com/privacy"
-
-}
