--- feature/vk-pay/src/main/java/com/appcoins/wallet/feature/vkpay/VkPayManager.kt
@@ -1,115 +1,0 @@
-package com.appcoins.wallet.feature.vkpay
-
-import android.app.Application
-import android.content.Context
-import androidx.appcompat.content.res.AppCompatResources
-import androidx.fragment.app.FragmentManager
-import com.vk.auth.main.VkClientUiInfo
-import com.vk.auth.main.VkSilentTokenExchanger
-import com.vk.dto.common.id.UserId
-import com.vk.superapp.SuperappKit
-import com.vk.superapp.SuperappKitConfig
-import com.vk.superapp.core.SuperappConfig
-import com.vk.superapp.vkpay.checkout.VkPayCheckout
-import com.vk.superapp.vkpay.checkout.api.dto.model.VkMerchantInfo
-import com.vk.superapp.vkpay.checkout.api.dto.model.VkTransactionInfo
-import com.vk.superapp.vkpay.checkout.config.VkPayCheckoutConfig
-import com.vk.superapp.vkpay.checkout.config.VkPayCheckoutConfigBuilder
-import com.vk.superapp.vkpay.checkout.data.VkCheckoutUserInfo
-import dagger.hilt.android.qualifiers.ApplicationContext
-import javax.inject.Inject
-
-
-class VkPayManager @Inject constructor(
-  @ApplicationContext private val context: Context,
-  private val walletVKSilentTokenExchanger: VkSilentTokenExchanger
-) {
-
-  private val application
-    get() = context as Application
-
-  fun initSuperAppKit(
-    appName: String,
-    clientSecret: String,
-    iconResources: Int,
-    vkSdkAppId: String,
-  ) {
-    if (!SuperappKit.isInitialized()) {
-      val icon = AppCompatResources.getDrawable(application, iconResources)!!
-      val appInfo = SuperappConfig.AppInfo(
-        appName = appName,
-        appId = vkSdkAppId,
-        appVersion = VKConstants.APP_VK_VERSION
-      )
-
-      val config = SuperappKitConfig.Builder(application)
-        .setAuthModelData(clientSecret)
-        .setAuthUiManagerData(VkClientUiInfo(icon, appName))
-        .setLegalInfoLinks(
-          serviceUserAgreement = VKConstants.SERVICE_USER_AGREEMENT,
-          servicePrivacyPolicy = VKConstants.SERVICE_PRIVACY_POLICY
-        )
-        .setApplicationInfo(appInfo)
-        .setUseCodeFlow(true)
-        .setSilentTokenExchanger(walletVKSilentTokenExchanger)
-        .disableMyTrackerAnalytics()
-        .build()
-
-      SuperappKit.init(config)
-    }
-  }
-
-  fun checkoutVkPay(
-    hash: String,
-    uidTransaction: String,
-    email: String,
-    phone: String,
-    walletAddress: String,
-    amount: Int,
-    vkMerchantId: Int,
-    vkSdkAppId: Int,
-    fragmentManager: FragmentManager
-  ) {
-    val transaction = VkTransactionInfo(
-      amount = amount,
-      orderId = createVKData(
-        uid = uidTransaction,
-        phone = phone,
-        email = email
-      ),
-      currency = VkTransactionInfo.Currency.RUB
-    )
-    val merchantInfo = VkMerchantInfo(
-      merchantId = vkMerchantId,
-      merchantSignature = hash,
-      merchantUserId = walletAddress,
-      merchantName = "wallet APPC"
-    )
-
-    //This Val need to implement only in Developer Mode
-    val config = if (BuildConfig.DEBUG) {
-      val sandbox = VkPayCheckoutConfig.Environment.Sandbox(
-        userInfo = VkCheckoutUserInfo(UserId(12345), "+1234566790"),
-        useApi = false,
-        mockNotCreatedVkPay = true,
-        useTestMerchant = true,
-        domain = VkPayCheckoutConfig.Domain.TEST
-      )
-      VkPayCheckoutConfigBuilder(merchantInfo)
-        .setParentAppId(vkSdkAppId)
-        .setEnvironment(sandbox)
-        .build()
-    } else {
-      VkPayCheckoutConfigBuilder(merchantInfo)
-        .setParentAppId(vkSdkAppId)
-        .build()
-    }
-    VkPayCheckout.startCheckout(
-      fm = fragmentManager,
-      transactionInfo = transaction,
-      config = config
-    )
-  }
-
-  private fun createVKData(uid: String, phone: String, email: String) = "$uid;$phone;$email"
-}
