+++ gradle.properties
@@ -45,8 +45,6 @@ INDICATIVE_API_KEY="INDICATIVE_API_KEY"
 INDICATIVE_API_KEY_DEV="INDICATIVE_API_KEY_DEV"
 AMPLITUDE_API_KEY="AMPLITUDE_API_KEY"
 FYBER_APP_ID="FYBER_APP_ID"
-VK_EXTERNAL_AUTH_REDIRECT_HOST=vk.com
-VK_CLIENT_SECRET_DEV="VK_CLIENT_SECRET_DEV"
 android.defaults.buildfeatures.buildconfig=true
 android.nonTransitiveRClass=false
 android.nonFinalResIds=false
