+++ settings.gradle.kts
@@ -23,17 +23,6 @@ dependencyResolutionManagement {
     maven { url = uri("https://s01.oss.sonatype.org/content/repositories/snapshots") }
     // needed for pincode Lollipin
     maven { url = uri("https://github.com/omadahealth/omada-nexus/raw/master/release") }
-    // VK ID and Vk Pay
-    maven {
-      url = uri("https://artifactory-external.vkpartner.ru/artifactory/superappkit-maven-public/")
-    }
-    maven {
-      credentials  {
-        username = "i.sharkov"
-        password = "R85uFTY8k16i"
-      }
-      url = uri("https://artifactory-external.vkpartner.ru/artifactory/superappkit-maven/")
-    }
   }
 }
